Refactored member routes
