+++ server/src/routes/members.routes.js
@@ -1,12 +1,16 @@
-/* import express from 'express';
+import express from 'express';
 import {
-  customerForm,
-  postCustomer,
-} from '../controller/customers.controller.js';
+  getMember,
+  postMember,
+  updateMember,
+  deleteMember,
+} from '../controller/members.controller.js';
 
 const router = express.Router();
 
-router.get(customerForm);
-router.post(postCustomer);
+router.get('/members', getMember);
+router.post('/members', postMember);
+router.put('/:memberId', updateMember);
+router.delete('/:memberId', deleteMember);
 
-export default router; */
+export default router;
