+++ server/src/server.js
@@ -4,6 +4,7 @@ import express from 'express';
 import mongoose from 'mongoose';
 
 import Member from './models/member.model.js';
+import memberRoutes from './routes/members.routes.js';
 
 const connectionString = 'mongodb://localhost:27017/remember-app';
 
@@ -11,9 +12,6 @@ mongoose.connect(connectionString, {
   useNewUrlParser: true,
   useUnifiedTopology: true,
 });
-
-/* import productRoutes from './routes/products.routes.js';
-import customerRoutes from './routes/customers.routes.js'; */
 
 const server = express();
 
@@ -24,56 +22,6 @@ server.get('/', (request, response) => {
   response.json({ status: 'Server is up and running' });
 });
 
-server.get('/members', (request, response) => {
-  Member.find()
-    .then((members) => response.json(members))
-    .catch((error) => response.json(error.message));
-});
-
-server.post('/members', (request, response) => {
-  const newMember = new Member({
-    firstName: request.body.firstName,
-    lastName: request.body.lastName,
-    description: request.body.description,
-    group: request.body.group,
-    image: request.body.image,
-    entries: request.body.entries,
-  });
-
-  newMember
-    .save()
-    .then((member) => response.json(member))
-    .catch((error) => response.json(error.message));
-});
-
-server.put('/:memberId', (request, response) => {
-  const memberId = request.params.memberId;
-
-  const updatedMember = {
-    firstName: request.body.firstName,
-    lastName: request.body.lastName,
-    description: request.body.description,
-    group: request.body.group,
-    image: request.body.image,
-    entries: request.body.entries,
-  };
-
-  Member.findOneAndUpdate({ _id: memberId }, updatedMember, {
-    new: true,
-  })
-    .then((member) => response.json(member))
-    .catch((error) => response.json(error.message));
-});
-
-server.delete('/:memberId', (request, response) => {
-  const idToDelete = request.params.memberId;
-
-  Member.findOneAndDelete({ _id: idToDelete }).catch((error) =>
-    response.json(error.message)
-  );
-});
-
-/* app.use(productRoutes);
-app.use(customerRoutes); */
+server.use(memberRoutes);
 
 server.listen(4000, () => console.log('Server started'));
